refactor: 사용X 변수 제거 및 임포트 수정
