+++ src/apis/axiosInstance.ts
@@ -1,5 +1,5 @@
 import axios, { AxiosInstance } from 'axios';
-import { getStorageItem } from '@src/utils/storage';
+import { getStorageItem } from '@utils/storage';
 
 const host = process.env.REACT_APP_API_ENDPOINT ?? 'http://localhost:3000';
