+++ src/layout/Layout.tsx
@@ -2,11 +2,11 @@ import { useLocation } from 'react-router-dom';
 import { Outlet } from 'react-router';
 import { useDispatch, useSelector } from 'react-redux';
 import { useEffect, useLayoutEffect } from 'react';
+import { getStorageItem } from '@utils/storage';
+import { loginUser, selectUser } from '@store/slices/user';
 import { selectFlashAlert } from '@store/slices/flashAlert';
-import { getStorageItem } from '@src/utils/storage';
-import { loginUser, selectUser } from '@src/store/slices/user';
-import { getAuthUser } from '@src/apis/user';
 import { FlashAlert, NavigationHeader } from '@containers';
+import { getAuthUser } from '@apis/user';
 
 import { LayoutContainer } from './style';
