+++ src/pages/NotFound/NotFound.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom';
-import LogoIcon from '@src/components/LogoIcon/LogoIcon';
 import { Button, Typography, useTheme } from '@mui/material';
+import { LogoIcon } from '@components';
 
 import { ButtonWrapper, NotFoundContainer } from './style';
